refactor: move style components to src/components
