+++ site/src/components/AccentClip.tsx
@@ -1,0 +1,16 @@
+import {styled} from '../stitches.config'
+
+export const AccentClip = styled('div', {
+  backgroundColor: '$accentSolid',
+  color: 'transparent',
+  variants: {
+    clip: {
+      text: {
+        backgroundClip: 'text',
+      },
+    },
+  },
+  defaultVariants: {
+    clip: 'text',
+  },
+})
